+++ s2gos-client/src/s2gos_client/__init__.py
@@ -7,7 +7,7 @@ from importlib.metadata import version
 from .api.async_client import AsyncClient
 from .api.client import Client
 from .api.config import ClientConfig
-from .api.exceptions import ClientException
+from .api.exceptions import ClientError
 
 __version__ = version("s2gos-client")
 
@@ -15,6 +15,6 @@ __all__ = [
     "AsyncClient",
     "Client",
     "ClientConfig",
-    "ClientException",
+    "ClientError",
     "__version__",
 ]