+++ s2gos-client/src/tests/cli/test_client.py
@@ -8,8 +8,9 @@ import click
 import pytest
 import typer
 
-from s2gos_client import ClientException
+from s2gos_client import ClientError
 from s2gos_client.api.client import Client
+from s2gos_client.api.transport import TransportError
 from s2gos_client.cli.cli import cli
 from s2gos_client.cli.client import use_client
 from s2gos_common.models import ApiError
@@ -22,20 +23,9 @@ class UseClientTest(TestCase):
 
     # noinspection PyMethodMayBeStatic
     def test_fail_with_client_error(self):
-        with pytest.raises(
-            click.ClickException,
-            match=(
-                "Not found\n"
-                "Server-side error details:\n"
-                "  title:  Not found\n"
-                "  status: 404\n"
-                "  type:   error\n"
-                "  detail: Something was not found\n"
-                "  traceback:\n"
-            ),
-        ):
-            with use_client(new_cli_context(traceback=True), None):
-                raise ClientException(
+        with pytest.raises(click.exceptions.Exit):
+            with use_client(new_cli_context(), None):
+                raise ClientError(
                     "Not found",
                     api_error=ApiError(
                         type="error",
@@ -47,7 +37,34 @@ class UseClientTest(TestCase):
                 )
 
     # noinspection PyMethodMayBeStatic
-    def test_fail_with_non_client_error(self):
+    def test_fail_with_client_error_and_traceback(self):
+        with pytest.raises(ClientError, match="Not found"):
+            with use_client(new_cli_context(traceback=True), None):
+                raise ClientError(
+                    "Not found",
+                    api_error=ApiError(
+                        type="error",
+                        title="Not found",
+                        status=404,
+                        detail="Something was not found",
+                        traceback=["a", "b", "c"],
+                    ),
+                )
+
+    # noinspection PyMethodMayBeStatic
+    def test_fail_with_transport_error(self):
+        with pytest.raises(click.exceptions.Exit):
+            with use_client(new_cli_context(), None):
+                raise TransportError("connection could not be established")
+
+    # noinspection PyMethodMayBeStatic
+    def test_fail_with_transport_error_and_traceback(self):
+        with pytest.raises(TransportError, match="connection could not be established"):
+            with use_client(new_cli_context(traceback=True), None):
+                raise TransportError("connection could not be established")
+
+    # noinspection PyMethodMayBeStatic
+    def test_fail_with_other_error(self):
         with pytest.raises(ValueError, match=r"path must be given"):
             with use_client(new_cli_context(), None):
                 raise ValueError("path must be given")