+++ CHANGES.md
@@ -1,6 +1,12 @@
 ## Changes in version 0.0.5 (in development)
 
 
+- In `s2gos_client`:
+  - Improved error handling (#54)
+  - Renamed `ClientException` into `ClientError`. 
+  - Renamed `TransportException` into `TransportError`. 
+
+
 ## Changes in version 0.0.4
 
 - Prevent server from logging `/jobs` request, as they are used for polling. 
@@ -12,6 +18,7 @@
 - Added a keyword-argument `inputs_arg: str | bool` to the `ProcessRegistry.process`
   decorator. If specified, it defines an _inputs argument_, which is used to 
   define the process inputs in form of a dataclass derived from 
+- 
   `pydantic.BaseModel`. (#35)
 - Now supporting the OpenAPI/JSON Schema `discriminator` property 
   in `tools/openapi.yaml/schemas/schema` and generated models. (#36)